+++ node/src/Commands.ts
@@ -942,9 +942,9 @@ export function createBLMove(
  * @internal
  */
 export function createLSet(
-    key: string,
+    key: GlideString,
     index: number,
-    element: string,
+    element: GlideString,
 ): command_request.Command {
     return createCommand(RequestType.LSet, [key, index.toString(), element]);
 }
@@ -953,7 +953,7 @@ export function createLSet(
  * @internal
  */
 export function createLTrim(
-    key: string,
+    key: GlideString,
     start: number,
     end: number,
 ): command_request.Command {
@@ -968,9 +968,9 @@ export function createLTrim(
  * @internal
  */
 export function createLRem(
-    key: string,
+    key: GlideString,
     count: number,
-    element: string,
+    element: GlideString,
 ): command_request.Command {
     return createCommand(RequestType.LRem, [key, count.toString(), element]);
 }
@@ -989,8 +989,8 @@ export function createRPush(
  * @internal
  */
 export function createRPushX(
-    key: string,
-    elements: string[],
+    key: GlideString,
+    elements: GlideString[],
 ): command_request.Command {
     return createCommand(RequestType.RPushX, [key].concat(elements));
 }
@@ -2239,12 +2239,16 @@ export function createBLPop(
  * @internal
  */
 export function createFCall(
-    func: string,
-    keys: string[],
-    args: string[],
+    func: GlideString,
+    keys: GlideString[],
+    args: GlideString[],
 ): command_request.Command {
-    let params: string[] = [];
-    params = params.concat(func, keys.length.toString(), keys, args);
+    const params: GlideString[] = [
+        func,
+        keys.length.toString(),
+        ...keys,
+        ...args,
+    ];
     return createCommand(RequestType.FCall, params);
 }
 
@@ -2252,12 +2256,16 @@ export function createFCall(
  * @internal
  */
 export function createFCallReadOnly(
-    func: string,
-    keys: string[],
-    args: string[],
+    func: GlideString,
+    keys: GlideString[],
+    args: GlideString[],
 ): command_request.Command {
-    let params: string[] = [];
-    params = params.concat(func, keys.length.toString(), keys, args);
+    const params: GlideString[] = [
+        func,
+        keys.length.toString(),
+        ...keys,
+        ...args,
+    ];
     return createCommand(RequestType.FCallReadOnly, params);
 }
 
@@ -2265,7 +2273,7 @@ export function createFCallReadOnly(
  * @internal
  */
 export function createFunctionDelete(
-    libraryCode: string,
+    libraryCode: GlideString,
 ): command_request.Command {
     return createCommand(RequestType.FunctionDelete, [libraryCode]);
 }
@@ -2285,7 +2293,7 @@ export function createFunctionFlush(mode?: FlushMode): command_request.Command {
  * @internal
  */
 export function createFunctionLoad(
-    libraryCode: string,
+    libraryCode: GlideString,
     replace?: boolean,
 ): command_request.Command {
     const args = replace ? ["REPLACE", libraryCode] : [libraryCode];
@@ -2295,7 +2303,7 @@ export function createFunctionLoad(
 /** Optional arguments for `FUNCTION LIST` command. */
 export type FunctionListOptions = {
     /** A wildcard pattern for matching library names. */
-    libNamePattern?: string;
+    libNamePattern?: GlideString;
     /** Specifies whether to request the library code from the server or not. */
     withCode?: boolean;
 };
@@ -2303,7 +2311,7 @@ export type FunctionListOptions = {
 /** Type of the response of `FUNCTION LIST` command. */
 export type FunctionListResponse = Record<
     string,
-    string | Record<string, string | string[]>[]
+    GlideString | Record<string, GlideString | GlideString[]>[]
 >[];
 
 /**
@@ -2312,7 +2320,7 @@ export type FunctionListResponse = Record<
 export function createFunctionList(
     options?: FunctionListOptions,
 ): command_request.Command {
-    const args: string[] = [];
+    const args: GlideString[] = [];
 
     if (options) {
         if (options.libNamePattern) {
@@ -2335,7 +2343,7 @@ export function createFunctionList(
 export type FunctionStatsSingleResponse = Record<
     string,
     | null
-    | Record<string, string | string[] | number> // Running function/script information
+    | Record<string, GlideString | GlideString[] | number> // Running function/script information
     | Record<string, Record<string, number>> // Execution engines information
 >;
 
@@ -3821,12 +3829,12 @@ export function createLMPop(
  * @internal
  */
 export function createBLMPop(
-    timeout: number,
-    keys: string[],
+    keys: GlideString[],
     direction: ListDirection,
+    timeout: number,
     count?: number,
 ): command_request.Command {
-    const args: string[] = [
+    const args: GlideString[] = [
         timeout.toString(),
         keys.length.toString(),
         ...keys,